[BUGFIX] Skip inactive pages when adding redirect
Since we now skip the redirect conflict check for inactive pages we must
not add potentially conflicting redirects either.

This is especially important when a page with a future "starttime" is
renamed back and forth; previously a redirect for each version was added
and no more conflict resolving was performed. This in turn lead to the
page not being accessible (too many redirects) once it was published.

See 4a81f61abfa833970492d2816d782316522d7681